Extract event wrap and unwrap code so it can be called from mlrun
[ML-7202](https://iguazio.atlassian.net/browse/ML-7202)
